+++ storage/src/remove_graph.cc
@@ -39,7 +39,7 @@ remove_graph::remove_graph()
  * @param index_id an index id
  * @param is_index a boolean telling if it is an index
  */
-remove_graph::remove_graph(uint32_t index_id, bool is_index)
+remove_graph::remove_graph(uint64_t index_id, bool is_index)
     : io::data(remove_graph::static_type()), id{index_id}, is_index{is_index} {}
 
 /**************************************
